requirements.txt
celery==4.3.0
channels==2.2.0
channels-redis==2.4.0
daphne==2.3.0
Django<3
django-bootstrap4==0.0.8
django-cleanup==3.2.0
django-cors-headers==3.0.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-yasg==1.15.0
fcm-django==0.2.21
gunicorn
opencv-python==4.1.0.25
Pillow==6.0.0
psycopg2-binary==2.8.2
redis==3.2.1
requests
sorl-thumbnail==12.5.0
uwsgi
